+++ game/ts/element-game.ts
@@ -14,6 +14,7 @@ import { endTreeCanvas, getElementTree, initTreeCanvas } from "./tree";
 import Color from "color";
 import { IsNullAPI } from "../../shared/api/internal/internal-null";
 import { elementPopAnimation, elementErrorAnimation } from "./element-game/element-animations"
+import { formatSuggestDisplay, updateSuggestion, ElementDom} from "./utils";
 
 function formatCategory(x: string) {
   return x.split('-').map(capitalize).join(' ')
@@ -49,7 +50,7 @@ let holdingRect: DOMRect;
 
 let suggestLeft: Elem;
 let suggestRight: Elem;
-let suggestResult: E4Suggestion;
+export let suggestResult: E4Suggestion;
 
 export function getElementMargin() {
   const x = document.querySelector('.elem') as HTMLElement;
@@ -99,32 +100,6 @@ async function dropHoldingElement(combineWith?: HTMLElement) {
       x.remove();
     }
   }
-}
-
-function formatSuggestDisplay(x: number) {
-  const y = Math.round(x * 50 + 50);
-  return y + '%';
-}
-
-function updateSuggestion() {
-  suggestResultElem.value = suggestResult.text;
-  suggestResultElem.setAttribute('style', getCSSFromDisplay({ text: '', color: `${[suggestResult.color.base,suggestResult.color.saturation,suggestResult.color.lightness].join('_')}` }));
-  document.querySelector('.color-btn.selected')?.classList.remove('selected');
-  document.querySelector(`.color-btn[data-color="${suggestResult.color.base}"]`).classList.add('selected');
-  (document.querySelector('[data-suggestion-slider="saturation"]') as HTMLInputElement).value = suggestResult.color.saturation.toString();
-  (document.querySelector('[data-suggestion-slider-value="saturation"]') as HTMLSpanElement).innerHTML = formatSuggestDisplay(suggestResult.color.saturation);
-  (document.querySelector('[data-suggestion-slider="lightness"]') as HTMLInputElement).value = suggestResult.color.lightness.toString();
-  (document.querySelector('[data-suggestion-slider-value="lightness"]') as HTMLSpanElement).innerHTML = formatSuggestDisplay(suggestResult.color.lightness);
-  (document.querySelector(`.suggest-hint`) as HTMLElement).style.color = suggestResultElem.style.color;
-}
-
-// Makes the HTML for an element
-export function ElementDom(elem2: Elem) {
-  const display = elem2.display;
-  const elem = document.createElement('div');
-  elem.appendChild(document.createTextNode(display.text));
-  elem.className = `elem ${getClassFromDisplay(display)}`;
-  return elem;
 }
 
 // Adds an element and has most element logic
